requirements/base.txt
Django==1.6.7
Pillow==2.6.1
South==1.0
bpython==0.13.1
django-braces==1.4.0
django-extensions==1.4.4
django-model-utils==2.0.3
logutils==0.3.3
#psycopg2==2.5.4
MySQL-python==1.2.5
django-auth-ldap==1.2.1
